# docs/swagger.yaml
basePath: /api/
definitions:
  handler.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.AdminCreateRentInput:
    properties:
      priceOfUnit:
        type: number
      priceType:
        type: string
      timeEnd:
        type: string
      timeStart:
        type: string
      transportId:
        type: integer
      userId:
        type: integer
    required:
    - priceOfUnit
    - priceType
    - timeStart
    - transportId
    - userId
    type: object
  models.AdminCreateRentOutput:
    properties:
      rent:
        $ref: '#/definitions/models.AdminGetRentOutput'
    type: object
...
  models.AdminEndRentInput:
    properties:
      lat:
        maximum: 180
        minimum: -180
        type: number
      long:
        maximum: 180
        minimum: -180
        type: number
    required:
    - lat
    - long
    type: object
  models.AdminEndRentOutput:
    properties:
      rent:
        $ref: '#/definitions/models.AdminGetRentOutput'
    type: object
...
  models.AdminGetRentOutput:
    properties:
      isDeleted:
        type: boolean
      rent:
        $ref: '#/definitions/models.GetRentOutput'
    type: object
  models.AdminGetTransportHistoryOutput:
    properties:
      rents:
        items:
          $ref: '#/definitions/models.AdminGetRentOutput'
        type: array
    type: object
...
  models.AdminGetUserHistoryOutput:
    properties:
      rents:
        items:
          $ref: '#/definitions/models.AdminGetRentOutput'
        type: array
    type: object
...
  models.AdminUpdateRentInput:
    properties:
      priceOfUnit:
        type: number
      priceType:
        type: string
      timeEnd:
        type: string
      timeStart:
        type: string
      transportId:
        type: integer
      userId:
        type: integer
    required:
    - priceOfUnit
    - priceType
    - timeStart
    - transportId
    - userId
    type: object
  models.AdminUpdateRentOutput:
    properties:
      rent:
        $ref: '#/definitions/models.AdminGetRentOutput'
    type: object
...
  models.GetAccountOutput:
    properties:
      balance:
        type: number
      id:
        type: integer
      isAdmin:
        type: boolean
      username:
        type: string
    type: object
  models.GetRentOutput:
    properties:
      ID:
        type: integer
      account:
        type: integer
      finalPrice:
        type: number
      priceType:
        type: string
      priceUnit:
        type: number
      timeEnd:
        type: string
      timeStart:
        type: string
      transport:
        type: integer
    type: object
...
  models.PaymentHesoyamUserOutput:
    properties:
      account:
        $ref: '#/definitions/models.GetAccountOutput'
    type: object
...
info:
  contact: {}
  description: API for transportation rental service
  title: SimbirGoAPI
  version: 1.0.0
paths:
...
  /Admin/Rent:
    post:
      consumes:
      - application/json
      description: Create new rent.
      parameters:
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AdminCreateRentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminCreateRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Create rent
      tags:
      - admin-rent
  /Admin/Rent/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting rental information by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Delete rent
      tags:
      - admin-rent
    get:
      consumes:
      - application/json
      description: Return rent information by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminGetRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Rent information
      tags:
      - admin-rent
    put:
      consumes:
      - application/json
      description: Changing a lease record by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdateRentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminUpdateRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Update rent
      tags:
      - admin-rent
  /Admin/Rent/End/{id}:
    post:
      consumes:
      - application/json
      description: Completion of the lease of transportation under the lease id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      - description: '-'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AdminEndRentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminEndRentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: End rent
      tags:
      - admin-rent
...
  /Admin/TransportHistory/{id}:
    get:
      consumes:
      - application/json
      description: Return transport rent history by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminGetTransportHistoryOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Transport rent history
      tags:
      - admin-rent
  /Admin/UserHistory/{id}:
    get:
      consumes:
      - application/json
      description: Return user rent history by id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminGetUserHistoryOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: User rent history
      tags:
      - admin-rent
  /Payment/Hesoyam/{id}:
    post:
      consumes:
      - application/json
      description: Adds 250,000 cash units to the balance of the account with id.
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentHesoyamUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - BearerAuth: []
      summary: Hesoyam
      tags:
      - payment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"